Added validation to tickets and exhibitions. Updated controllers

+++ src/main/java/com/Kiiko/ExhibitionsApp/controller/TicketController.java
@@ -1,12 +1,17 @@
 package com.Kiiko.ExhibitionsApp.controller;
 
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
+
 import com.Kiiko.ExhibitionsApp.api.TicketAPI;
 import com.Kiiko.ExhibitionsApp.controller.assembler.TicketAssembler;
 import com.Kiiko.ExhibitionsApp.controller.model.TicketModel;
 import com.Kiiko.ExhibitionsApp.dto.TicketDto;
+import com.Kiiko.ExhibitionsApp.exceptions.NoAccessToRecourseException;
 import com.Kiiko.ExhibitionsApp.service.TicketService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.hateoas.CollectionModel;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -23,16 +28,19 @@ public class TicketController implements TicketAPI {
     @Override
     public TicketModel getTicketInfo(Long userId, Long ticketId) {
         log.info("TicketController@getTicketInfo. Ticket ID : {}", ticketId);
-        //TODO: check if user is owner of this ticket
+        if (!ticketService.isTicketBelongsToUser(ticketId, userId)) {
+            throw new NoAccessToRecourseException();
+        }
         return ticketAssembler.toModel(ticketService.getTicketDetails(ticketId));
     }
 
     @Override
-    public List<TicketModel> getUsersTickets(Long userId) {
+    public CollectionModel<TicketModel> getUsersTickets(Long userId) {
         log.info("TicketController@getUsersTickets. User ID : {}", userId);
-        return ticketService.getUsersTickets(userId).stream()
+        List<TicketModel> ticketList = ticketService.getUsersTickets(userId).stream()
                 .map(ticketAssembler::toModel)
                 .collect(Collectors.toList());
+        return CollectionModel.of(ticketList, linkTo(methodOn(TicketController.class).getUsersTickets(userId)).withSelfRel());
     }
 
     @Override
@@ -43,8 +51,11 @@ public class TicketController implements TicketAPI {
 
     @Override
     public ResponseEntity<Void> deleteTicket(Long userId, Long ticketId) {
-        //TODO: check if user is owner of this ticket
         log.info("TicketController@deleteTicket. Ticket ID : {}", ticketId);
+        if (!ticketService.isTicketBelongsToUser(ticketId, userId)) {
+            log.info("Access to ticket rejected {}", ticketId);
+            throw new NoAccessToRecourseException();
+        }
         ticketService.deleteTicket(ticketId);
         return ResponseEntity.noContent().build();
     }
